+++ changelog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.0.6] [Unreleased] [Github Repo]
+## Added
+- Non-blocking cmd_nb method
+- methods and decorators to parse for/while loops and functions
+- decorator to define a function as a shortcut to a function in MicroPython
 
 ## [0.0.5] - 2019-10-21
 ### Added